+++ library/src/main/java/com/lazy/library/logging/Config.java
@@ -29,6 +29,7 @@ class Config {
         this.logCatLogLevel = builder.logCatLogLevel;
         this.fileLogLevel = builder.fileLogLevel;
         this.topLevelTag = builder.topLevelTag;
+        this.jLog = builder.jLog;
         this.autoSaveLogToFile = builder.autoSaveLogToFile;
         this.showStackTraceInfo = builder.showStackTraceInfo;
         this.showFileTimeInfo = builder.showFileTimeInfo;