+++ changelog.md
@@ -1,6 +1,14 @@
 
 # changelog
 
+# v2.1.1
+
++ Custom File Log Format
++ deleteUnusedLogEntriesAfterDays
++ Add File Tag
+* Custom top-level switch
+* Log StackTraceInfo switch
+
 # v2.0.0
 
 + add JLog